refactor in order to expose executeMutateOffline
removed the OfflineFirstRelay file, now the store is created in the environment's constructor

created the public function "executeMutateOffline" in the environment and used by the environment when the "executeMutate" is performed and the application is offline

type refactors, they are now all managed in the RelayOfflineTypes file and exported to the index file

fix SSR EnvionmentIDB
