+++ packages/relay-offline/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@wora/relay-offline",
-  "version": "2.0.4",
+  "version": "2.1.0-rc.1",
   "keywords": [
     "wora",
     "cache-persist",
