+++ packages/relay-offline/src/EnvironmentIDB.ts
@@ -1,13 +1,10 @@
 import { ICacheStorage, CacheOptions } from '@wora/cache-persist';
 import IDBStorage, { IOnUpgrade } from '@wora/cache-persist/lib/idbstorage';
-import { EnvironmentConfig } from 'relay-runtime/lib/store/RelayModernEnvironment';
 import { Store, RecordSource } from '@wora/relay-store';
 import { Scheduler, OperationLoader } from 'relay-runtime/lib/store/RelayStoreTypes';
 import RelayModernEnvironment from './RelayModernEnvironment';
 import { CacheOptionsStore } from '@wora/relay-store/lib/Store';
-
-type Omit<T, K extends keyof T> = Pick<T, Exclude<keyof T, K>>;
-type EnvironmentOfflineConfig = Omit<EnvironmentConfig, 'store'>; // Equivalent to: {b: number, c: boolean}
+import { EnvironmentOfflineConfig } from './RelayOfflineTypes';
 
 class EnvironmentIDB {
     public static create(
@@ -26,10 +23,22 @@ class EnvironmentIDB {
         } = {},
         offlineStoreOptions: CacheOptions = {},
     ): RelayModernEnvironment {
-        let idbStore: CacheOptions;
-        let idbRecords: CacheOptions;
-        let idbOffline: CacheOptions;
         const { gcScheduler, operationLoader, getDataID, persistOptions } = storeOptions;
+        const idbStore: CacheOptions = {
+            serialize: false,
+            prefix: null,
+            ...persistOptions,
+        };
+        const idbRecords: CacheOptions = {
+            serialize: false,
+            prefix: null,
+            ...recordSourceOptions,
+        };
+        const idbOffline: CacheOptions = {
+            serialize: false,
+            prefix: null,
+            ...offlineStoreOptions,
+        };
         if (typeof window !== 'undefined') {
             const { name = 'relay', onUpgrade, version } = idbOptions;
             const idbStorages: ICacheStorage[] = IDBStorage.create({
@@ -39,26 +48,9 @@ class EnvironmentIDB {
                 version,
             });
 
-            idbStore = {
-                storage: idbStorages[0],
-                serialize: false,
-                prefix: null,
-                ...persistOptions,
-            };
-
-            idbRecords = {
-                storage: idbStorages[1],
-                serialize: false,
-                prefix: null,
-                ...recordSourceOptions,
-            };
-
-            idbOffline = {
-                storage: idbStorages[2],
-                serialize: false,
-                prefix: null,
-                ...offlineStoreOptions,
-            };
+            idbStore.storage = idbStorages[0];
+            idbRecords.storage = idbStorages[1];
+            idbOffline.storage = idbStorages[2];
         }
         const recordSource = new RecordSource(idbRecords);
         const store = new Store(recordSource, idbStore, gcScheduler, operationLoader, getDataID);
