+++ packages/relay-offline/src/index.ts
@@ -1,4 +1,4 @@
 export { default as Environment } from './RelayModernEnvironment';
-export { default as OfflineStore } from './OfflineFirstRelay';
 export { Store } from '@wora/relay-store';
 export { RecordSource } from '@wora/relay-store';
+export * from './RelayOfflineTypes';
